changes to accompany those in https://github.com/lookit/lookit-api/tree/feature/62-pattern-preview-after-participate to allow saving data and using child info during preview
