+++ app/controllers/participate.js
@@ -3,7 +3,6 @@ import Ember from 'ember';
 // Adapted from Lookit's participate controller https://github.com/CenterForOpenScience/lookit/blob/develop/app/controllers/participate.js
 export default Ember.Controller.extend({
     isDirty: function() {
-        // TODO: check the response model to see if it contains unsaved data
         var response = this.get('response');
         return response.get('hasDirtyAttributes');
     },
