+++ app/controllers/preview.js
@@ -1,34 +1,5 @@
-import Ember from 'ember';
+import Participate from './participate';
 
-// Adapted from Experimenter's preview controller https://github.com/CenterForOpenScience/experimenter/blob/develop/app/controllers/experiments/info/preview.js
-export default Ember.Controller.extend({
-    breadCrumb: 'Preview',
-    experiment: null,
-    session: null,
-    isDirty: function() {
-        return this.get('model.hasDirtyAttributes');
-    },
-
-    _resolve: null,
-    previewData: null,
-    showData: false,
-    showPreviewData(session) {
-        return new Ember.RSVP.Promise((resolve) => {
-            this.set('previewData', JSON.stringify(session.toJSON(), null, 4));
-            this.set('showData', true);
-            this.set('_resolve', resolve);
-        });
-    },
-    actions: {
-        toggleData() {
-            this.toggleProperty('showData');
-            this.get('_resolve')();
-        },
-        saveResponse(payload, callback) {
-            var response = this.get('response');
-            response.setProperties(payload);
-            response.save().then(callback);
-            this.set('response', response);
-        }
-    }
-});
+// Just use the participation controller directly; placeholder in case we later want to
+// have slightly different behavior for the preview
+export default Participate;
