+++ app/routes/preview.js
@@ -1,33 +1,12 @@
-import Ember from 'ember';
-import WarnOnExitRouteMixin from '../mixins/warn-on-exit-route';
-import FramePlayerRoute from '../mixins/frame-player-route';
+import Participate from './participate';
 
-// Adapted from Experimenter preview route https://github.com/CenterForOpenScience/experimenter/blob/develop/app/routes/experiments/info/preview.js
-export default Ember.Route.extend(WarnOnExitRouteMixin, FramePlayerRoute, {
+// Adapted from Lookit participate route https://github.com/CenterForOpenScience/lookit/blob/develop/app/routes/participate.js
+export default Participate.extend({
     _createStudyResponse() {
         const response = this._super();
         response.setProperties({
-            id: 'PREVIEW_DATA_DISREGARD'
+            isPreview: true
         });
-        return response.reopen({
-            save() {
-                if (this.get('completed')) {
-                    var controller = Ember.getOwner(this).lookup('controller:preview');
-                    controller.showPreviewData(this).then(() => {
-                        // Override the WarnOnExitMixin's behavior
-                        controller.set('forceExit', true);
-                        window.location.href = '/exp/studies/';
-                    });
-                    return Ember.RSVP.reject();
-                } else {
-                    return Ember.RSVP.resolve(this);
-                }
-            }});
-    },
-    _getChild() {
-        let child = this.store.createRecord('child', {
-            id: 'TEST_CHILD_DISREGARD'
-        });
-        return child;
+        return response;
     }
 });
